improved functionality of show courses
add, deletes courses and returns to main

+++ GradeCalculator/app/src/main/java/com/example/kevin/gradecalculator/DBHelper.java
@@ -246,7 +246,7 @@ public class DBHelper extends SQLiteOpenHelper {
         return courseDistribution;
     }
     //Deleters
-    public boolean deleteCourseById(int id) {
+    public boolean deleteCourseById(long id) {
         // obtain a database connection
         SQLiteDatabase database = this.getWritableDatabase();
